Higher-level custom op API, V3
This PR introduces:
- a FunctonalBaseOp class.
- To define a new custom op, a user subclasses FunctionalBaseOp, adds
  their device-specific implementations (by adding static methods to
  their subclass), and adds an abstract impl by overriding the
  `abstract` staticmethod.
- Under the hood, we take the class and all its methods and call the
  corresponding torch.library.{define, impl} APIs to register it.

The class-based approach more closely resembles PyTorch C++ codegen;
we get all of the information about the operator up front in the class
and we can do things with this information, like add an "autograd not
implemented" kernel if the user didn't specify an autograd kernel (NB:
this functionality is not in this PR).

Please see the docstrings for example usages.

ghstack-source-id: ad36532c5a34bb71cdd4ed1135172dbeb0615310
Pull Request resolved: https://github.com/pytorch/pytorch/pull/111661